requirements.txt
astroid==2.0.2
certifi==2018.10.15
chardet==3.0.4
Django==2.1
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pygments==2.2.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.5
requests==2.19.1
selenium==3.14.0
six==1.11.0
typed-ast==1.1.0
typing==3.6.4
urllib3==1.23
wrapt==1.10.11
